+++ feature/notetaking/src/main/java/com/iguana/notetaking/ai/AiViewModel.kt
@@ -1,15 +1,15 @@
-package com.iguana.notetaking
+package com.iguana.notetaking.ai
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class RecordViewModel: ViewModel() {
+class AiViewModel : ViewModel() {
+
     private val _pageNumber = MutableLiveData<Int>()
     val pageNumber: LiveData<Int> get() = _pageNumber
 
     fun setPageNumber(pageNumber: Int) {
         _pageNumber.value = pageNumber
     }
-
 }
