+++ feature/notetaking/src/main/java/com/iguana/notetaking/recording/RecordViewModel.kt
@@ -1,15 +1,15 @@
-package com.iguana.notetaking
+package com.iguana.notetaking.recording
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class AiViewModel : ViewModel() {
-
+class RecordViewModel: ViewModel() {
     private val _pageNumber = MutableLiveData<Int>()
     val pageNumber: LiveData<Int> get() = _pageNumber
 
     fun setPageNumber(pageNumber: Int) {
         _pageNumber.value = pageNumber
     }
+
 }
