+++ feature/notetaking/src/main/res/layout/fragment_ai.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.iguana.notetaking.AiViewModel" />
+            type="com.iguana.notetaking.ai.AiViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
