+++ feature/notetaking/src/main/res/layout/fragment_record.xml
@@ -4,7 +4,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.iguana.notetaking.RecordViewModel" />
+            type="com.iguana.notetaking.recording.RecordViewModel" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
